# src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Banking Products API
servers:
  - url: http://localhost:8585
tags:
  - name: Consultar
  - name: Crear
  - name: Transferir
paths:
  /products/{id}:
    get:
      summary: Obtener los datos de un producto por id
      operationId: getProductById
      tags:
        - Consultar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID del cliente
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        '404':
          description: Producto no encontrado
  /products/saving-accounts:
    post:
      summary: Crea una cuenta de ahorro
      tags:
        - Crear
      operationId: createSavingAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavingAccountRequest"
      responses:
        '201':
          description: Cuenta de ahorro creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /products/current-accounts:
    post:
      summary: Crea una cuenta de corriente
      tags:
        - Crear
      operationId: createCurrentAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCurrentAccountRequest"
      responses:
        '201':
          description: Cuenta de corriente creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /products/fixed-term-deposits:
    post:
      summary: Crea un plazo fijo
      tags:
        - Crear
      operationId: createFixedTermDeposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFixedTermDepositRequest"
      responses:
        '201':
          description: Plazo fijo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /products/transfer-money:
    post:
      summary: Transferir dinero desde un producto a otro
      tags:
        - Transferir
      operationId: transferMoney
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferMoneyRequest"
      responses:
        '204':
          description: Se ha realizado la transferencia correctamente
        '400':
          description: Error de validación o parámetros incorrectos enviados por el cliente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor. Ha ocurrido un problema inesperado en el microservicio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Descripción del error generado
        status:
          type: string
          description: Clasificación del error
          example: BAD_REQUEST
      required:
        - message
        - status
    CreateSavingAccountRequest:
      type: object
      required:
        - clientId
      properties:
        clientId:
          type: string
    CreateCurrentAccountRequest:
      type: object
      required:
        - clientId
      properties:
        clientId:
          type: string
    CreateFixedTermDepositRequest:
      type: object
      required:
        - clientId
        - balance
        - maturityDate
      properties:
        clientId:
          type: string
        balance:
          type: number
          format: double
          multipleOf: 0.01
          minimum: 500
        maturityDate:
          type: string
          format: date
    AccountResponse:
      type: object
      required:
        - id
        - accountNumber
        - clientId
        - balance
        - openedDate
        - isActive
        - type
        - maintenceFee
        - maturityDate
        - interestRate
        - currentWithdrawalCount
      properties:
        id:
          type: string
        accountNumber:
          type: string
        clientId:
          type: string
        balance:
          type: number
          format: double
          multipleOf: 0.01
        openedDate:
          type: string
          format: date
        active:
          type: boolean
        type:
          type: string
          enum: [ SAVING_ACCOUNT, CURRENT_ACCOUNT, FIXED_TERM_DEPOSIT ]
        maintenceFee:
          type: number
          format: double
          multipleOf: 0.01
        maturityDate:
          type: string
          format: date
        interestRate:
          type: number
          format: double
          multipleOf: 0.01
        currentWithdrawalCount:
          type: integer
    TransferMoneyRequest:
      type: object
      required:
        - amount
        - productOriginId
        - productDestinationId
      properties:
        amount:
          type: number
          format: double
          multipleOf: 0.01
          description: Monto a transferir, con hasta dos decimales (por ejemplo, 100.50).
        productOriginId:
          type: string
          description: Identificador del producto bancario de origen.
        productDestinationId:
          type: string
          description: Identificador del producto bancario de destino.